requirements.txt
algoliasearch
annotated-types
anyio
appnope
asttokens
Brotli
certifi
cffi
charset-normalizer
comm
configparser
cryptography
debugpy
decorator
distro
dnspython
exceptiongroup
executing
googleapis-common-protos
grpc-gateway-protoc-gen-openapiv2
grpcio
h11
httpcore
httpx
idna
importlib-metadata
ipykernel
ipython
jedi
jiter
jupyter_client
jupyter_core
loguru
lz4
matplotlib-inline
nest-asyncio
numpy
openai
packaging
parso
pexpect
pickleshare
pinecone-client
platformdirs
prompt-toolkit
protobuf
psutil
ptyprocess
pure-eval
pybase64
pycparser
pydantic
pydantic_core
Pygments
pyOpenSSL
PySocks
python-dateutil
python-dotenv
PyYAML
pyzmq
requests
setuptools
six
sniffio
stack-data
tornado
tqdm
traitlets
typing_extensions
urllib3
wcwidth
wheel
zipp
